requirements.txt
django==4.2
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2-binary==2.9.6
djoser==2.2.0
django-debug-toolbar==4.1.0
pytest==7.1.2
pytest-django==4.5.2
coverage==7.2.6
pytest-cov==4.1.0
django-filter==23.2
yookassa==2.3.6
python-dotenv==1.0.0
flake8==6.0.0
celery==5.3.1
redis==4.5.5
flower==2.0.0
pytest-celery==0.0.0